Primitives (#11)

+++ packages/geometry/include/geometry/Vector3.h
@@ -13,17 +13,17 @@ public:
 	/**
 	 * X coordinate
 	 */
-	double x;
+	float x;
 
 	/**
 	 * Y coordinate
 	 */
-	double y;
+	float y;
 
 	/**
 	 * Z coordinate
 	 */
-	double z;
+	float z;
 
 	/**
 	 * Constructs a vector where all coordinate values are zero
@@ -36,7 +36,7 @@ public:
      * @param y Y coordinate value
 	 * @param z Z coordinate value
 	 */
-	Vector3(double x, double y, double z);
+	Vector3(float x, float y, float z);
 
 	/**
 	 * Constructs a vector using the x and y coordinates from a two-dimensional vector
@@ -55,7 +55,7 @@ public:
      * @param tol Tolerance to use when comparing. Defaults to 0.
      * @return True if this vector is the zero vector. Returns false otherwise.
      */
-    bool isZero(double tol = 0.0) const;
+    bool isZero(float tol = 0.0) const;
 
     /**
      * Determines if this vector is equal to another vector (within tolerance).
@@ -63,19 +63,19 @@ public:
      * @param tol Tolerance value to use when comparing. Defaults to 0.
      * @return True if the two vectors are equal. Returns false otherwise.
      */
-    bool equalTo(const Vector3& other, double tol = 0.0) const;
+    bool equalTo(const Vector3& other, float tol = 0.0) const;
 
     /**
      * @return The magnitude of this vector
      */
-    double getMagnitude() const;
+    float getMagnitude() const;
 
     /**
      * Computes the dot product of this vector with another vector
      * @param other The other vector
      * @return The dot product of this vector and the given vector
      */
-    double dot(const Vector3& other) const;
+    float dot(const Vector3& other) const;
 
     /**
      * Computes the cross product of this vector with another vector.
@@ -98,7 +98,7 @@ public:
      * @param result (Optional) Pointer to a vector in which to store the results
      * @return The resulting vector
      */
-    Vector3 scale(double value, Vector3* result = nullptr) const;
+    Vector3 scale(float value, Vector3* result = nullptr) const;
 
     /**
      * Subtracts another vector from this vector.
